test(e2e): share the waits, fixtures, and teardown across the browser suites
One helper now owns the text/overlay waits (all on interval polling —
only one of the five copies had received the hidden-page fix), the app
fixtures (accepted, unaccepted, warning-carrying, boom), the overlay
and indicator element ids, and the browser/app teardown that every
describe had cloned. Two snapshots move a warning column, following
the shared fixture's indentation.
