+++ test/e2e/logging.test.js
@@ -1,36 +1,14 @@
 import collectConsole, { normalizeConsole } from "../helpers/console-collector";
+import {
+  acceptedApp,
+  closeE2e,
+  waitForAppText,
+  warningApp,
+} from "../helpers/e2e";
 import createHotApp from "../helpers/hot-app";
 import runBrowser from "../helpers/run-browser";
 
 jest.setTimeout(400000);
-
-/**
- * @param {string} text text rendered into #app
- * @returns {string} app source
- */
-function app(text) {
-  return `
-    document.getElementById("app").textContent = ${JSON.stringify(text)};
-    if (module.hot) {
-      module.hot.accept();
-    }
-  `;
-}
-
-/**
- * @param {import("puppeteer").Page} page page
- * @param {string} text expected #app text
- * @returns {Promise<void>} resolved when rendered
- */
-function waitForAppText(page, text) {
-  return page
-    .waitForFunction(
-      (expected) => document.getElementById("app")?.textContent === expected,
-      { timeout: 30000 },
-      text,
-    )
-    .then(() => {});
-}
 
 describe("client logging (browser)", () => {
   let hotApp;
@@ -38,24 +16,11 @@ describe("client logging (browser)", () => {
   let page;
 
   afterEach(async () => {
-    // try/finally: a rejected browser.close() must not leak the watcher,
-    // the server, and the temp dir behind it.
-    try {
-      if (browser) {
-        await browser.close();
-      }
-    } finally {
-      browser = undefined;
-      if (hotApp) {
-        const closing = hotApp;
-        hotApp = undefined;
-        await closing.close();
-      }
-    }
+    ({ browser, app: hotApp } = await closeE2e(browser, hotApp));
   });
 
   it("logs a full update cycle at the default info level, with the prefix", async () => {
-    hotApp = await createHotApp({ code: app("v1") });
+    hotApp = await createHotApp({ code: acceptedApp("v1") });
     ({ page, browser } = await runBrowser());
     const console_ = collectConsole(page);
 
@@ -63,7 +28,7 @@ describe("client logging (browser)", () => {
     await waitForAppText(page, "v1");
     await console_.waitFor("connected");
 
-    hotApp.edit(app("v2"));
+    hotApp.edit(acceptedApp("v2"));
     await waitForAppText(page, "v2");
     await console_.waitFor("App is up to date");
 
@@ -71,7 +36,10 @@ describe("client logging (browser)", () => {
   });
 
   it("logging=log adds the collapsed per-module detail", async () => {
-    hotApp = await createHotApp({ query: "?logging=log", code: app("v1") });
+    hotApp = await createHotApp({
+      query: "?logging=log",
+      code: acceptedApp("v1"),
+    });
     ({ page, browser } = await runBrowser());
     const console_ = collectConsole(page);
 
@@ -79,7 +47,7 @@ describe("client logging (browser)", () => {
     await waitForAppText(page, "v1");
     await console_.waitFor("connected");
 
-    hotApp.edit(app("v2"));
+    hotApp.edit(acceptedApp("v2"));
     await waitForAppText(page, "v2");
     await console_.waitFor("App is up to date");
 
@@ -89,14 +57,17 @@ describe("client logging (browser)", () => {
   });
 
   it("logging=none silences the whole cycle", async () => {
-    hotApp = await createHotApp({ query: "?logging=none", code: app("v1") });
+    hotApp = await createHotApp({
+      query: "?logging=none",
+      code: acceptedApp("v1"),
+    });
     ({ page, browser } = await runBrowser());
     const console_ = collectConsole(page);
 
     await page.goto(hotApp.url);
     await waitForAppText(page, "v1");
 
-    hotApp.edit(app("v2"));
+    hotApp.edit(acceptedApp("v2"));
     await waitForAppText(page, "v2");
 
     // The update applied (asserted through the DOM above) — give any stray
@@ -113,18 +84,7 @@ describe("client logging (browser)", () => {
       query: "?logging=warn",
       // `require(<expression>)` produces webpack's "Critical dependency"
       // warning without failing the build.
-      code: `
-        document.getElementById("app").textContent = "v1";
-        const dep = "./nothing";
-        try {
-          require(dep);
-        } catch (err) {
-          // expected
-        }
-        if (module.hot) {
-          module.hot.accept();
-        }
-      `,
+      code: warningApp("v1"),
     });
     ({ page, browser } = await runBrowser());
     const console_ = collectConsole(page);
@@ -138,7 +98,10 @@ describe("client logging (browser)", () => {
   });
 
   it("logging=error keeps errors only", async () => {
-    hotApp = await createHotApp({ query: "?logging=error", code: app("v1") });
+    hotApp = await createHotApp({
+      query: "?logging=error",
+      code: acceptedApp("v1"),
+    });
     ({ page, browser } = await runBrowser());
     const console_ = collectConsole(page);
